+++ src/exceptions.c
@@ -31,6 +31,9 @@ PERFIDIOUS_PUBLIC zend_class_entry *perfidious_pmu_not_found_exception_ce;
 PERFIDIOUS_PUBLIC zend_class_entry *perfidious_pmu_event_not_found_exception_ce;
 PERFIDIOUS_PUBLIC zend_class_entry *perfidious_overflow_exception_ce;
 PERFIDIOUS_PUBLIC zend_class_entry *perfidious_io_exception_ce;
+PERFIDIOUS_PUBLIC zend_class_entry *perfidious_closed_exception_ce;
+PERFIDIOUS_PUBLIC zend_class_entry *perfidious_wrong_thread_exception_ce;
+PERFIDIOUS_PUBLIC zend_class_entry *perfidious_resource_busy_exception_ce;
 PERFIDIOUS_PUBLIC zend_class_entry *perfidious_unsupported_metric_exception_ce;
 
 // clang-format off
@@ -128,6 +131,51 @@ static zend_class_entry *register_class_IOException(zend_class_entry *restrict i
 
 PERFIDIOUS_ATTR_RETURNS_NONNULL
 PERFIDIOUS_ATTR_WARN_UNUSED_RESULT
+static zend_class_entry *register_class_ClosedException(zend_class_entry *restrict iface)
+{
+    zend_class_entry ce;
+    zend_class_entry *class_entry;
+
+    INIT_CLASS_ENTRY(ce, PHP_PERFIDIOUS_NAMESPACE "\\ClosedException", NULL);
+    class_entry = zend_register_internal_class_ex(&ce, spl_ce_LogicException);
+    class_entry->ce_flags |= ZEND_ACC_FINAL | ZEND_ACC_NO_DYNAMIC_PROPERTIES;
+    zend_class_implements(class_entry, 1, iface);
+
+    return class_entry;
+}
+
+PERFIDIOUS_ATTR_RETURNS_NONNULL
+PERFIDIOUS_ATTR_WARN_UNUSED_RESULT
+static zend_class_entry *register_class_WrongThreadException(zend_class_entry *restrict iface)
+{
+    zend_class_entry ce;
+    zend_class_entry *class_entry;
+
+    INIT_CLASS_ENTRY(ce, PHP_PERFIDIOUS_NAMESPACE "\\WrongThreadException", NULL);
+    class_entry = zend_register_internal_class_ex(&ce, spl_ce_LogicException);
+    class_entry->ce_flags |= ZEND_ACC_FINAL | ZEND_ACC_NO_DYNAMIC_PROPERTIES;
+    zend_class_implements(class_entry, 1, iface);
+
+    return class_entry;
+}
+
+PERFIDIOUS_ATTR_RETURNS_NONNULL
+PERFIDIOUS_ATTR_WARN_UNUSED_RESULT
+static zend_class_entry *register_class_ResourceBusyException(zend_class_entry *restrict iface)
+{
+    zend_class_entry ce;
+    zend_class_entry *class_entry;
+
+    INIT_CLASS_ENTRY(ce, PHP_PERFIDIOUS_NAMESPACE "\\ResourceBusyException", NULL);
+    class_entry = zend_register_internal_class_ex(&ce, spl_ce_RuntimeException);
+    class_entry->ce_flags |= ZEND_ACC_FINAL | ZEND_ACC_NO_DYNAMIC_PROPERTIES;
+    zend_class_implements(class_entry, 1, iface);
+
+    return class_entry;
+}
+
+PERFIDIOUS_ATTR_RETURNS_NONNULL
+PERFIDIOUS_ATTR_WARN_UNUSED_RESULT
 static zend_class_entry *register_class_UnsupportedMetricException(zend_class_entry *restrict iface)
 {
     zend_class_entry ce;
@@ -178,6 +226,9 @@ void perfidious_exceptions_minit(void)
     perfidious_pmu_event_not_found_exception_ce =
         register_class_PmuEventNotFoundException(perfidious_exception_interface_ce);
     perfidious_io_exception_ce = register_class_IOException(perfidious_exception_interface_ce);
+    perfidious_closed_exception_ce = register_class_ClosedException(perfidious_exception_interface_ce);
+    perfidious_wrong_thread_exception_ce = register_class_WrongThreadException(perfidious_exception_interface_ce);
+    perfidious_resource_busy_exception_ce = register_class_ResourceBusyException(perfidious_exception_interface_ce);
     perfidious_unsupported_metric_exception_ce =
         register_class_UnsupportedMetricException(perfidious_exception_interface_ce);
 }